% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_themes.R
\name{a_concert_theme}
\alias{a_concert_theme}
\title{All of the themes follow a similar pattern mainly just changing out font families
and small tweaks to plot elements}
\usage{
a_concert_theme(base_family = "Inconsolata", base_size = 10,
  plot_title_family = "Concert One", plot_title_size = 25,
  plot_title_face = "bold", subtitle_family = base_family,
  subtitle_size = 10, subtitle_face = "plain",
  strip_text_family = base_family, strip_text_size = 12,
  strip_text_face = "plain", caption_family = base_family,
  caption_size = 10, caption_face = "plain",
  axis_text_size = base_size, axis_title_family = plot_title_family,
  axis_title_size = 12, axis_title_face = "plain",
  grid_col = "#dedede", grid = TRUE, noaxis = FALSE,
  emphasis = "xy", rotate_x = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size}{plot title family, face, size}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size}{plot caption family, face, size}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{grid_col, axis_col}{grid & axis colors; both default to \code{#cccccc}}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{noaxis}{include/exclude axis (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{emphasis}{bold one axis title of both (a combination of 'x','y' or 'xy')}

\item{rotate_x}{TRUE/FAlSE to rotate axis text}
}
\description{
a_theme / a_concert_theme / a_robot_theme are all mostly unused now
}
