% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{a_pal}
\alias{a_pal}
\title{Original color palette with decent brightness}
\usage{
a_pal()
}
\description{
Original color palette with decent brightness
}
\examples{
library(scales)
scales::show_col(a_pal()(8))
}
