% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{col_compare}
\alias{col_compare}
\title{Function to return the column with the max value (or value of function provided)}
\usage{
col_compare(df, cols, fun = which.max)
}
\arguments{
\item{df}{dataframe object}

\item{cols}{column(s) to remove NA and replace with 0}

\item{fun}{operator function to apply across columns, defaults to max}
}
\description{
Function to return the column with the max value (or value of function provided)
}
\keyword{max}
\keyword{value}
