% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{m.compress}
\alias{m.compress}
\title{for use with axis labels turns big number into abbrev. form}
\usage{
m.compress(tx)
}
\description{
for use with axis labels turns big number into abbrev. form
}
