% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{p_pal}
\alias{p_pal}
\title{Primary color palette with decent brightness}
\usage{
p_pal()
}
\description{
Primary color palette with decent brightness
}
\examples{
library(scales)
scales::show_col(p_pal()(8))
}
