% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_gt.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{crop_gt}
\title{Pipe operator}
\usage{
crop_gt(file, whitespace)

lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
The result of calling `rhs(lhs)`.
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
