% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ppalette}
\alias{ppalette}
\title{primary color palette}
\format{An object of class \code{character} of length 8.}
\usage{
ppalette
}
\description{
primary color palette
}
\keyword{datasets}
