% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{removeNAs}
\alias{removeNAs}
\title{Remove NAs function}
\usage{
removeNAs(df, cols)
}
\arguments{
\item{df}{dataframe object}

\item{cols}{column(s) to remove NA and replace with 0}
}
\description{
Remove NAs function
}
\keyword{NA}
