% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{spalette}
\alias{spalette}
\title{secondary color palette}
\format{An object of class \code{character} of length 6.}
\usage{
spalette
}
\description{
secondary color palette
}
\keyword{datasets}
